[ao] Removing memoryless observer args for MovingAverage
The original implementation of memoryless observers used MinMaxObservers and
a memoryless argument to manipulate the behavior of the observer such that it wouldn't
keep track of previously observed min and max's. It was later pointed
out that this was equivalent to a movingaverageobserver with averaging_constant=1
which is requires less overhead and no 1 off args (memoryless) so this PR refactors
the memoryless arg and uses MovingAverage observers instead, although the memoryless
adjective is still used, a complete definintion was also added to clarify error
messages given these changes.

ghstack-source-id: de8ba7edd1fb04e067804dca20828bc5db20962a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73947